# CullPilot/docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: CullPilot 后端接口
  version: 1.0.0
  description: CullPilot 图片筛选系统的业务接口。Apifox 请求统一访问 Java 服务，Java 服务再调用 Python AI API。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - 用户认证
      summary: 注册用户
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: Password123!
              displayName: 图片筛选用户
      responses:
        '201':
          description: 注册成功并自动登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseEnvelope'
        '409':
          description: 邮箱已经注册
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 注册参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags:
        - 用户认证
      summary: 用户登录
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: Password123!
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseEnvelope'
        '401':
          description: 邮箱或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags:
        - 用户认证
      summary: 获取当前用户
      operationId: getCurrentUser
      responses:
        '200':
          description: 当前用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseEnvelope'
        '401':
          description: 未登录或 Token 已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags:
        - 用户认证
      summary: 退出登录
      operationId: logoutUser
      responses:
        '200':
          description: 退出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponseEnvelope'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/projects/{projectId}/assets:
    post:
      tags:
        - 图片资源
      summary: 批量上传图片
      operationId: uploadAssets
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: binary
                  description: 图片文件，支持 JPEG、PNG、WebP，单个文件最大 20 MB。可一次选择多个文件。
                file:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: 兼容旧客户端的文件字段；通常使用 files。
      responses:
        '201':
          description: 上传完成，允许部分成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadBatchResponseEnvelope'
        '404':
          description: 项目不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: 文件超过大小限制
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: 不支持的图片格式
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 整批文件均上传失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 图片资源
      summary: 查询项目图片列表
      operationId: listProjectAssets
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: decision
          in: query
          required: false
          description: 用户决策筛选。
          schema:
            type: string
            enum: [keep, review, reject]
        - name: recommendation
          in: query
          required: false
          description: 系统推荐筛选。
          schema:
            type: string
            enum: [keep, review, reject]
        - name: groupId
          in: query
          required: false
          description: 按分组 ID 筛选。
          schema:
            type: string
            format: uuid
        - name: analysisStatus
          in: query
          required: false
          description: 按分析状态筛选。
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: sort
          in: query
          required: false
          description: 排序表达式，支持 createdAt、updatedAt、sizeBytes、originalName。
          schema:
            type: string
            default: createdAt:desc
            example: createdAt:asc
      responses:
        '200':
          description: 图片分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPageResponseEnvelope'
        '404':
          description: 项目不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/assets/{assetId}:
    get:
      tags:
        - 图片资源
      summary: 查询图片详情
      operationId: getAsset
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: 图片详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponseEnvelope'
        '404':
          description: 图片不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/assets/{assetId}/decision:
    patch:
      tags:
        - 图片资源
      summary: Update user decision
      operationId: updateAssetDecision
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDecisionRequest'
      responses:
        '200':
          description: Decision updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponseEnvelope'
        '404':
          description: Asset not found or not owned by current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/assets/{assetId}/thumbnail:
    get:
      tags:
        - 图片资源
      summary: 获取图片缩略图
      operationId: getAssetThumbnail
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: JPEG 缩略图二进制流
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: 图片或缩略图不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/assets/{assetId}/original:
    get:
      tags:
        - 图片资源
      summary: 获取原图
      operationId: getAssetOriginal
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: 原图二进制流
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          description: 图片或原图不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/projects/{projectId}/export:
    post:
      tags:
        - 图片导出
      summary: 创建导出任务
      operationId: createExport
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportRequest'
            example:
              selection: keepAndReview
              copyImages: true
              stripGps: true
              includeManifest: true
              manifestFormat: csv
      responses:
        '202':
          description: 导出任务已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponseEnvelope'
        '404':
          description: 项目不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 导出参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/exports/{exportId}:
    get:
      tags:
        - 图片导出
      summary: 查询导出任务
      operationId: getExport
      parameters:
        - $ref: '#/components/parameters/ExportId'
      responses:
        '200':
          description: 导出任务状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponseEnvelope'
        '404':
          description: 导出任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/exports/{exportId}/download:
    get:
      tags:
        - 图片导出
      summary: 下载导出压缩包
      operationId: downloadExport
      parameters:
        - $ref: '#/components/parameters/ExportId'
      responses:
        '200':
          description: 导出 ZIP 文件
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          description: 导出任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 导出任务尚未完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: 项目 UUID。
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    AssetId:
      name: assetId
      in: path
      required: true
      description: 图片资源 UUID。
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440001
...
    ExportId:
      name: exportId
      in: path
      required: true
      description: 导出任务 UUID
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440003
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 相同项目内用于安全重试启动请求的幂等键。
      schema:
        type: string
        maxLength: 200
    Page:
      name: page
      in: query
      required: false
      description: 页码，从 1 开始。
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: 每页返回的数据条数。
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 50
  schemas:
    AnalysisStatus:
      type: string
      enum:
        - pending
        - processing
        - completed
        - failed
    AssetDecision:
      type: string
      enum:
        - keep
        - review
        - reject
    AssetRecommendation:
      type: string
      nullable: true
      enum:
        - keep
        - review
        - reject
...
    Asset:
      type: object
      required:
        - id
        - projectId
        - originalName
        - mimeType
        - sizeBytes
        - width
        - height
        - sha256
        - analysisStatus
        - decision
        - recommendation
        - thumbnailUrl
        - originalUrl
        - exif
        - quality
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        originalName:
          type: string
        mimeType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
        sizeBytes:
          type: integer
          format: int64
        width:
          type: integer
        height:
          type: integer
        sha256:
          type: string
          description: 原图内容的 SHA-256 哈希。
        analysisStatus:
          $ref: '#/components/schemas/AnalysisStatus'
        decision:
          $ref: '#/components/schemas/AssetDecision'
        recommendation:
          $ref: '#/components/schemas/AssetRecommendation'
        groupId:
          type: string
          format: uuid
          nullable: true
        rank:
          type: integer
          nullable: true
          description: 组内 AI 排名，从 1 开始
        recommendScore:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
        recommendReasons:
          type: array
          items: {type: string}
        membershipReason:
          type: string
          nullable: true
        features:
          type: object
          description: 可读的技术特征摘要，不包含原始图片路径或人脸向量
        thumbnailUrl:
          type: string
        originalUrl:
          type: string
        exif:
          type: object
          description: EXIF 摘要，当前为空对象，后续分析模块补充。
        quality:
          type: object
          description: 图片质量结果，当前为空对象，后续分析模块补充。
        version:
          type: integer
          format: int64
          description: Current optimistic-lock version
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateDecisionRequest:
      type: object
      additionalProperties: false
      required:
        - decision
        - expectedVersion
      properties:
        decision:
          $ref: '#/components/schemas/AssetDecision'
        expectedVersion:
          type: integer
          format: int64
          minimum: 0
          description: Version returned by the last asset query
    AcceptedAsset:
      type: object
      required:
        - id
        - originalName
        - mimeType
        - sizeBytes
        - width
        - height
        - analysisStatus
        - decision
        - thumbnailUrl
        - originalUrl
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        width:
          type: integer
        height:
          type: integer
        analysisStatus:
          $ref: '#/components/schemas/AnalysisStatus'
        decision:
          $ref: '#/components/schemas/AssetDecision'
        recommendation:
          $ref: '#/components/schemas/AssetRecommendation'
        thumbnailUrl:
          type: string
        originalUrl:
          type: string
    DuplicateAsset:
      type: object
      required:
        - originalName
        - sha256
        - existingAssetId
      properties:
        originalName:
          type: string
        sha256:
          type: string
        existingAssetId:
          type: string
          format: uuid
    FailedAsset:
      type: object
      required:
        - originalName
        - code
        - message
      properties:
        originalName:
          type: string
        code:
          type: string
        message:
          type: string
    UploadBatch:
      type: object
      required:
        - batchId
        - accepted
        - duplicates
        - failed
      properties:
        batchId:
          type: string
          format: uuid
        accepted:
          type: array
          items:
            $ref: '#/components/schemas/AcceptedAsset'
        duplicates:
          type: array
          items:
            $ref: '#/components/schemas/DuplicateAsset'
        failed:
          type: array
          items:
            $ref: '#/components/schemas/FailedAsset'
    AssetPage:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
    AssetResponseEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Asset'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    AssetPageResponseEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/AssetPage'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    UploadBatchResponseEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/UploadBatch'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    CreateExportRequest:
      type: object
      additionalProperties: false
      required:
        - selection
      properties:
        selection:
          type: string
          enum: [keep, keepAndReview]
        copyImages:
          type: boolean
          default: true
        stripGps:
          type: boolean
          description: 指定时覆盖项目隐私设置；重新编码会移除所有 EXIF 元数据
        includeManifest:
          type: boolean
          default: true
        manifestFormat:
          type: string
          enum: [csv]
          default: csv
    ExportResponse:
      type: object
      required:
        - id
        - projectId
        - status
        - selection
        - progress
        - processedCount
        - totalCount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        selection:
          type: string
          enum: [keep, keepAndReview]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        processedCount:
          type: integer
        totalCount:
          type: integer
        errorMessage:
          type: string
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
    ExportResponseEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/ExportResponse'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    ApiMeta:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
          format: uuid
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
        displayName:
          type: string
          maxLength: 100
          nullable: true
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    User:
      type: object
      required:
        - id
        - email
        - displayName
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        status:
          type: string
          enum:
            - active
            - disabled
        createdAt:
          type: string
          format: date-time
    Auth:
      type: object
      required:
        - user
        - tokenType
        - accessToken
        - expiresAt
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokenType:
          type: string
          example: Bearer
        accessToken:
          type: string
          description: '登录凭证，后续请求放入 Authorization: Bearer <accessToken>。'
        expiresAt:
          type: string
          format: date-time
    AuthResponseEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Auth'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    UserResponseEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/User'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    Logout:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
    LogoutResponseEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Logout'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: PROJECT_NOT_FOUND
        message:
          type: string
          example: Project not found
        details:
          nullable: true
          description: 错误详细信息，具体结构取决于错误编码。
    ErrorResponse:
      type: object
      required:
        - error
        - meta
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ApiMeta'